requirements.txt
psycopg2-binary
flask
flask-cors
